feat: test utils and small optimization of Subs
